+++ CHANGELOG.md
@@ -12,6 +12,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Fixed
 - In Site Health section, Opcache status may be wrong (or generates PHP warnings) if OPcache API usage is restricted.
 
+### Removed
+- Analytics data collection when OPcache API usage is restricted.
+
 ## [2.1.0] - 2020-11-23
 
 ### Added